Outsourcing stac downlaoder to external backage for decoupling and reusability in other projects. Adding Microsft Planatary Computer as an additional imagery download source, to avoid data gaps in EarthSearch Collection 1
