+++ vercye_ops/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.1.dev93+g7c09e6f.d20250509'
+__version__ = '0.1.dev142+g8cd49c8.d20250630'
